Implementa o cancelamento de um projeto

+++ src/app/projetos/adicionar-kits/adicionar-kits.component.ts
@@ -19,8 +19,6 @@ export class AdicionarKitsComponent implements OnInit {
     kits: Kit[]=[];
     projetoId:number;
     projeto = Projeto.EMPTY_MODEL;
-    canEdit = false;
-    errorMessage = '';
 
     constructor(
         private kitDataService:KitDataService,
@@ -32,24 +30,11 @@ export class AdicionarKitsComponent implements OnInit {
 
     ngOnInit() {
         this.route.params.subscribe((params:Params)=>{
-            this.projetoId = params["id"];
+            this.projetoId = +params["id"];
             this.projetoDataService.getProjeto(this.projetoId).subscribe((projeto:Projeto)=>{
                 this.projeto = projeto;
                 this.kit_id = this.projeto.kit_id;
             });
-            this.projetoDataService.getEtapaEnvio(this.projetoId).subscribe(
-                (etapa:Etapa)=>{
-                    if(etapa && etapa.tarefas.length){
-                        console.log('não pode');
-                        this.errorMessage = "Não é possível fazer alterações nesta página. Já existe um plano de envio para este projeto.";
-                    }else{
-                        this.canEdit = true;
-                    }
-                }, error => {
-                    this.canEdit = true;
-                    console.log(error);
-                }
-            );
         });
 
         this.kitDataService.getKits().subscribe((kits:Kit[])=>{
@@ -59,12 +44,15 @@ export class AdicionarKitsComponent implements OnInit {
 
     onSubmit(form: NgForm) {
         let kitIdSelecionado = form.value.kit_id;
-        if(this.projeto.kit_id){
-            this.projetoDataService.deleteKit(this.projetoId, this.projeto.kit_id).subscribe(res=>{
-                this.projetoDataService.storeKit(this.projetoId, kitIdSelecionado).subscribe(res => {
-                    this.router.navigate(["../etapa-envio"], { relativeTo: this.route });
+        if(this.projeto.kit_id) {
+
+            if(kitIdSelecionado != this.projeto.kit_id && confirm("Ao modificar o kit de um projeto você estará cancelando as etapas subsequentes. Deseja continuar?")) {
+                this.projetoDataService.deleteKit(this.projetoId, this.projeto.kit_id).subscribe(res=>{
+                    this.projetoDataService.storeKit(this.projetoId, kitIdSelecionado).subscribe(res => {
+                        this.router.navigate(["../etapa-envio"], { relativeTo: this.route });
+                    });
                 });
-            });
+            }
         }else{
             this.projetoDataService.storeKit(this.projetoId, kitIdSelecionado).subscribe(res=>{
                 this.router.navigate(["../etapa-envio"], { relativeTo: this.route });
